style-component 적용
