+++ frontend/package.json
@@ -20,6 +20,7 @@
     "react-dom": "^19.1.0",
     "react-router-dom": "^7.5.0",
     "react-scripts": "5.0.1",
+    "styled-components": "^6.1.17",
     "typescript": "^5.8.3",
     "web-vitals": "^2.1.4"
   },
