+++ frontend/src/pages/serveypie/pages/CompletionPage/index.tsx
@@ -1,0 +1,12 @@
+import styled from "styled-components";
+
+function CompletionPage() {
+  return <CompletionPageContainer>CompletionPage</CompletionPageContainer>;
+}
+
+const CompletionPageContainer = styled.div`
+  background: aqua;
+  padding: 4em;
+`;
+
+export default CompletionPage;
