--- frontend/src/pages/serveypie/pages/PageA/index.tsx
@@ -1,5 +1,0 @@
-function PageA() {
-    return <div>PageA</div>
-}
-
-export default PageA;
