--- frontend/src/pages/serveypie/pages/PageB/index.tsx
@@ -1,5 +1,0 @@
-function PageB() {
-  return <div>PageB</div>;
-}
-
-export default PageB;
